refactor: move order APIs from /api/admin/orders to /api/orders for shared access
