--- backend/src/main/java/com/foodapp/controllers/AdminOrderController.java
@@ -1,63 +1,0 @@
-package com.foodapp.controllers;
-
-import com.foodapp.dto.requests.OrderRequest;
-import com.foodapp.dto.response.ApiResponse;
-import com.foodapp.services.Impl.OrderService;
-import jakarta.validation.Valid;
-import lombok.AccessLevel;
-import lombok.RequiredArgsConstructor;
-import lombok.experimental.FieldDefaults;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequestMapping("/api/admin/orders")
-@RequiredArgsConstructor
-@FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
-public class AdminOrderController {
-
-    OrderService orderService;
-
-    @GetMapping("/{userId}")
-    public ResponseEntity<?> getOrdersByUserId(@PathVariable Long userId) {
-        var orders = orderService.getOrdersByUserId(userId);
-        return ResponseEntity.ok(
-                ApiResponse.builder()
-                        .status(ApiResponse.Status.SUCCESS)
-                        .data(orders)
-                        .build()
-        );
-    }
-
-    @PostMapping
-    public ResponseEntity<?> createOrder(@Valid @RequestBody OrderRequest orderRequest) {
-        var orderResponse = orderService.createOrder(orderRequest);
-        return ResponseEntity.ok(
-                ApiResponse.builder()
-                        .status(ApiResponse.Status.SUCCESS)
-                        .data(orderResponse)
-                        .build()
-        );
-    }
-
-    @PutMapping("/{orderId}")
-    public ResponseEntity<?> updateOrder(@PathVariable Long orderId, @Valid @RequestBody OrderRequest orderRequest) {
-        var updatedOrder = orderService.updateOrder(orderId, orderRequest);
-        return ResponseEntity.ok(
-                ApiResponse.builder()
-                        .status(ApiResponse.Status.SUCCESS)
-                        .data(updatedOrder)
-                        .build()
-        );
-    }
-
-    @DeleteMapping("/{orderId}")
-    public ResponseEntity<?> deleteOrder(@PathVariable Long orderId) {
-        orderService.deleteOrder(orderId);
-        return ResponseEntity.ok(
-                ApiResponse.builder()
-                        .status(ApiResponse.Status.SUCCESS)
-                        .build()
-        );
-    }
-}
